Move AST, context and formatter to Debug namespace
Also fix remaining psalm errors.
